Fix read_html_live example
I checked the read_html_live example and saw that the css selectors changed and a cookie consent banner was added.

This PR is to changed the read_html_live() example, so it can reject cookies and extract organizations with the new page version. Scroll was needed to force the JSON file download.

I used |>, but I can change my PR to %>% if required.